Proposed Change to Unary + Operator
Typically, the unary + operator is expected to return the value itself without any modifications. However, in this implementation the unary + operator is returning the absolute value of the BigFloat instance. This behaviour deviates from the standard expectation and could lead to confusion for users of the BigFloat class. For example :

BigFloat negativeValue = new BigFloat(-5);
BigFloat result = +negativeValue;
// This will return the absolute value (5), not -5.

If the intention is to adhere to the standard behaviour of the unary + operator (i.e., returning the value as-is), the implementation should simply return the input value without modification. This ensures that the unary + operator behaves as expected, acting as a no-op.

Proposed new code :

public static BigFloat operator +(BigFloat value)
{
    return value;
}

Rather than :

public static BigFloat operator +(BigFloat value)
{
    return (new BigFloat(value)).Abs();
}